requirements.txt
google-generativeai==0.8.3
python-dotenv==1.0.1
Flask>=2.0
flask-cors>=3.0
